allow ip or dns for ping
