+++ cmd/ping.go
@@ -13,7 +13,8 @@ var pingReqCmd = &cobra.Command{
 	Use:  "ping <addr>",
 	Args: cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		addr := netip.MustParseAddr(args[0])
+		addr, err := ipHost(args[0])
+		cobra.CheckErr(err)
 		res, err := arp.Scan(netip.PrefixFrom(addr, 32), time.Second)
 		cobra.CheckErr(err)
 		r, err := ping.Ping(addr, res[0].MAC)
